Moved files around to reduce API Visibility of internal classes.
