+++ Tests/DataPipeline/EventPolicy/EventPolicyEngineTests.swift
@@ -1,4 +1,5 @@
-@testable import CioInternalCommon
+@testable import CioDataPipelines
+import CioInternalCommon
 import Foundation
 import SyncSqlCipher
 import Testing
